% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dom_optimal_allocation.R
\name{dom_optimal_allocation}
\alias{dom_optimal_allocation}
\title{Optimal sample size allocation}
\usage{
dom_optimal_allocation(
  id,
  Dom,
  H,
  Y,
  Rh = NULL,
  deffh = NULL,
  indicator,
  sup_w,
  sup_cv,
  min_size = 3,
  correction_before = FALSE,
  dataset = NULL
)
}
\arguments{
\item{id}{Variable for unit ID codes. One dimensional object convertible to one-column \code{data.table} or variable name as character, column number.}

\item{Dom}{Optional variables used to define population domains. If supplied, values are calculated for each domain. An object convertible to \code{data.table} or variable names as character vector, column numbers.}

\item{H}{The unit stratum variable. One dimensional object convertible to one-column \code{data.table} or variable name as character, column number.}

\item{Y}{Variable of interest. Object convertible to \code{data.table} or variable names as character, column numbers.}

\item{Rh}{The expected response rate in each stratum (optional). If not defined, it is assumed to be 1 in each stratum (full-response). Object convertible to one-column \code{data.table}, variable name as character, or column number.}

\item{deffh}{The expected design effect for the estimate of variable (optional). If not defined, it is assumed to be 1 for each variable in each stratum. If is defined, then variables is defined the same arrangement as \code{Yh}. Object convertible to \code{data.table}, variable name as character vector, or column numbers.}

\item{indicator}{Variable for detection fully surveyed units. Object convertible to \code{data.table} or variable names as character, column numbers.}

\item{sup_w}{Variable for weight limit in domain of stratum. Object convertible to \code{data.table} or variable names as character, column numbers.}

\item{sup_cv}{Variable for maximum coefficient of variation (CV) in percentage for domain. Object convertible to \code{data.table} or variable names as character, column numbers.}

\item{min_size}{A numeric value for sample size.}

\item{correction_before}{by default FALSE; correction of sample size is made before ending, if true, correction of sample size is made at the end.}

\item{dataset}{Optional survey data object convertible to \code{data.table} with one row for each stratum.}
}
\value{
A list with eights data objects:\cr
\item{data}{An object as \code{data.table}, with variables:
  \code{id} - variable with unit ID codes, \cr
  \code{Dom} - optional variables used to define population domains, \cr
  \code{H} - the unit stratum variable, \cr
  \code{Y} - variable of interest, \cr
  \code{Rh} - the expected response rate in each stratum, \cr
  \code{deffh} - the expected design effect, \cr
  \code{indicator} - variable for full surveys, \cr
  \code{sup_w} - variable for weight limit in domain of stratum, \cr
  \code{sup_cv} - Variable for maximum coefficient of variation, \cr
  \code{poph} - population size, \cr
  \code{nh} - sample size . }

\item{nh_larger_then_Nh}{An object as \code{data.table}, with variables: \cr
  \code{H} - the stratum variable, \cr
  \code{nh} - sample size,
  \code{poph} - population size.}

\item{dom_strata_size}{An object as \code{data.table}, with variables: \cr
  \code{H} - the unit stratum variable, \cr
  \code{Dom} - optional variables used to define population domains, \cr
  \code{sup_w} - variable for weight limit in domain of stratum, \cr
  \code{poph} - population size, \cr
  \code{nh} - sample size, \cr
  \code{sample100} - sample size for fully surveyed units, \cr
  \code{design_weights} - design weights. }

\item{dom_size}{An object as \code{data.table}, with variables: \cr
  \code{Dom} - optional variables used to define population domains, \cr
  \code{poph} - population size, \cr
  \code{nh} - sample size, \cr
  \code{sample100} - sample size for fully surveyed units, \cr
  \code{design_weights} - design weights. }

\item{size}{An object as \code{data.table}, with variables: \cr
  \code{poph} - population size, \cr
  \code{nh} - sample size, \cr
  \code{sample100} - sample size for fully surveyed units. }

\item{dom_strata_expected_precision}{An object as \code{data.table}, with variables: \cr
  \code{H} - stratum, \cr
  \code{variable} - the name of variable of interest, \cr
  \code{estim} - total value, \cr
  \code{deffh} - the expected design effect, \cr
  \code{s2h} - population variance \eqn{S^2}, \cr
  \code{nh} - sample size, \cr
  \code{Rh} - the expected response rate, \cr
  \code{deffh} - the expected design effect, \cr
  \code{poph} - population size, \cr
  \code{nrh} - expected number of respondents, \cr
  \code{var} - expected variance, \cr
  \code{se} - expected standard error, \cr
  \code{cv} - expected coefficient of variance.}

\item{dom_expected_precision}{An object as \code{data.table}, with variables: \cr
  \code{Dom} - domain, \cr
  \code{variable} - the name of variable of interest, \cr
  \code{poph} - the population size, \cr
  \code{nh} - sample size, \cr
  \code{nrh} - expected number of respondents, \cr
  \code{estim} - total value, \cr
  \code{var} - the expected variance, \cr
  \code{se} - the expected standard error, \cr
  \code{cv} - the expected coefficient of variance.}

\item{ total_expected_precision}{An object as \code{data.table}, with variables: \cr
  \code{variable} - the name of variable of interest, \cr
  \code{poph} - the population size, \cr
  \code{nh} - sample size, \cr
  \code{nrh} - expected number of respondents, \cr
  \code{estim} - total value, \cr
  \code{var} - the expected variance, \cr
  \code{se} - the expected standard error, \cr
  \code{cv} - the expected coefficient of variance.}
}
\description{
The function computes optimal sample size allocation over strata and domain for population.
}
\examples{
library("laeken")
library("data.table")
data("ses")
data <- data.table(ses)
data[, H := paste(location, NACE1, size, sep = "_")]
data[, id := .I]
data[, full := 0]
data[, sup_cv := 10]
data[, sup_w := 20]
#vars <- dom_optimal_allocation(id = "id", dom = "sex",
#                                H = "H", Y = "earnings",
#                                indicator = "full",
#                                sup_w = "sup_w",
#                                sup_cv = "sup_cv",
#                                min_size = 3,
#                                correction_before = FALSE,
#                                dataset = data)
#                                dataset=data)
#vars
}
\seealso{
\code{\link{expsize}}, \code{\link{optsize}}, \code{\link{prop_dom_optimal_allocation}}
}
\keyword{surveysampling}
