% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round2.R
\name{round2}
\alias{round2}
\title{Rounding numbers}
\usage{
round2(x, n)
}
\arguments{
\item{x}{a numeric vector.}

\item{n}{integer indicating the number of decimal places.}
}
\value{
Rounded value
}
\description{
The function rounds the values in its first argument to the specified number of decimal places (default 0).
}
\examples{
dar <- 100 * runif(3)
dar
round2(dar, 1)

}
\seealso{
\code{\link{expsize}}, \code{\link{dom_optimal_allocation}}
}
\keyword{surveysampling}
