\name{surveyplanning-package}
\alias{surveyplanning-package}
\alias{surveyplanning}
\docType{package}
\title{
Survey Planning Tools
}
\description{
Tools for sample survey planning, including sample size calculation,
estimation of expected precision for the estimates of totals, and
calculation of optimal sample size allocation.
}
\details{

Index:
\preformatted{
dom_optimal_allocation  Optimal sample size allocation
expsize                 Sample size calculation
expvar                  Expected precision for the estimates of totals
min_count               Minimal count of respondents for the given relative margin of error
min_prop                Minimal proportion for the given relative margin of error
MoE_Y                   Margin of error for count
MoE_P                   Margin of error for proportion
optsize                 Optimal sample size allocation
s2                      Population variance estimation
surveyplanning-package  Survey Planning Tools

}
}
\author{
Juris Breidaks [aut],
Martins Liberts [aut],
Janis Jukams [aut]
}
\keyword{ package }
