\name{surveyplanning-package}
\alias{surveyplanning-package}
\alias{surveyplanning}
\docType{package}
\title{
Survey Planning Tools
}
\description{
Tools for sample survey planning, including sample size calculation,
estimation of expected precision for the estimates of totals, and
calculation of optimal sample size allocation.
}
\details{
\tabular{ll}{
Package: \tab surveyplanning\cr
Version: \tab 2.9\cr
Date: \tab 2017-10-26\cr
Depends: \tab R (>= 3.0.0), data.table (>= 1.10.4), stats, laeken \cr
License: \tab GPL (>= 2)\cr
URL: \tab https://github.com/CSBLatvia/surveyplanning/ \cr
BugReports: \tab https://github.com/CSBLatvia/surveyplanning/issues/ \cr
}

Index:
\preformatted{
dom_optimal_allocation  Optimal sample size allocation
expsize                 Sample size calculation
expvar                  Expected precision for the estimates of totals
min_count               Minimal count of respondents for the given relative margin of error
min_prop                Minimal proportion for the given relative margin of error
MoE_Y                   Margin of error for count
MoE_P                   Margin of error for proportion
optsize                 Optimal sample size allocation
s2                      Population variance estimation
surveyplanning-package  Survey Planning Tools

}
}
\author{
Juris Breidaks [aut, cre],
Martins Liberts [aut],
Janis Jukams [aut]

Maintainer: Juris Breidaks <rcsb@csb.gov.lv>
}
\keyword{ package }
