\name{surveyplanning-package}
\alias{surveyplanning-package}
\alias{surveyplanning}
\docType{package}
\title{
Survey Planning tools
}
\description{
    Expexted size for sample, 
    expexted variance for variables of interest,
    optimal sample size for variables of interest.
}
\details{
\tabular{ll}{
Package: \tab surveyplanning\cr
Version: \tab 0.1\cr
Date: \tab 2015-10-29\cr
Depends: \tab R (>= 3.0.0), data.table (>= 1.9.6), laeken, stringr\cr
License: \tab GPL (>= 2)\cr
URL: \tab https://github.com/CSBLatvia/vardpoor/ \cr
BugReports: \tab https://github.com/CSBLatvia/vardpoor/issues/ \cr
}

Index:
\preformatted{
expsize	Expexted size for sample
expvar	Expexted variance for variables of interest
optsize	Optimal sample size for variables of interest
surveyplanning-package      Survey Planning tools

}
}
\author{
Juris Breidaks [aut, cre],
 Janis Jukams [aut]

Maintainer: Juris Breidaks <Juris.Breidaks@csb.gov.lv>
}
\keyword{ package }
