\name{surveyplanning-package}
\alias{surveyplanning-package}
\alias{surveyplanning}
\docType{package}
\title{
Survey Planning Tools
}
\description{
Tools for sample survey planing, including sample size calculation,
estimation of expected precision for the estimates of totals, and
calculation of optimal sample size allocation.
}
\details{
\tabular{ll}{
Package: \tab surveyplanning\cr
Version: \tab 0.2\cr
Date: \tab 2015-11-02\cr
Depends: \tab R (>= 3.0.0), data.table (>= 1.9.6), laeken, stringr\cr
License: \tab GPL (>= 2)\cr
URL: \tab https://github.com/CSBLatvia/surveyplanning/ \cr
BugReports: \tab https://github.com/CSBLatvia/surveyplanning/issues/ \cr
}

Index:
\preformatted{
expsize                 Sample size calculation
expvar                  Expected precision for the estimates of totals
optsize                 Optimal sample size allocation
s2                      Population variance estimation
surveyplanning-package  Survey Planning Tools

}
}
\author{
Juris Breidaks [aut, cre],
 Janis Jukams [aut]

Maintainer: Juris Breidaks <Juris.Breidaks@csb.gov.lv>
}
\keyword{ package }
